Also translate branch-metadata, introducing the new version of it.
