+++ crates/but-api/src/commands/workspace.rs
@@ -28,7 +28,7 @@ fn ref_metadata_toml(project: &Project) -> anyhow::Result<VirtualBranchesTomlMet
 #[api_cmd]
 #[tauri::command(async)]
 #[instrument(err(Debug))]
-pub fn head_info(project_id: ProjectId) -> Result<but_workspace::ui::ref_info::RefInfo, Error> {
+pub fn head_info(project_id: ProjectId) -> Result<but_workspace::ui::RefInfo, Error> {
     let project = gitbutler_project::get(project_id)?;
     let ctx = CommandContext::open(&project, AppSettings::load_from_default_path_creating()?)?;
     let repo = ctx.gix_repo_for_merging_non_persisting()?;
@@ -42,7 +42,7 @@ pub fn head_info(project_id: ProjectId) -> Result<but_workspace::ui::ref_info::R
         },
     )
     .map_err(Into::into)
-    .and_then(|info| but_workspace::ui::ref_info::RefInfo::for_ui(info, &repo).map_err(Into::into))
+    .and_then(|info| but_workspace::ui::RefInfo::for_ui(info, &repo).map_err(Into::into))
 }
 
 #[api_cmd]
